Make noCache and onlyIfCached internal. Also split ClientCacheExtensions into several files
