+++ normalized-cache/src/commonMain/kotlin/com/apollographql/cache/normalized/options/doNotStore.kt
@@ -1,0 +1,27 @@
+@file:Suppress("PackageDirectoryMismatch")
+
+package com.apollographql.cache.normalized
+
+import com.apollographql.apollo.api.ApolloRequest
+import com.apollographql.apollo.api.ExecutionContext
+import com.apollographql.apollo.api.MutableExecutionOptions
+import com.apollographql.apollo.api.Operation
+
+internal class DoNotStoreContext(val value: Boolean) : ExecutionContext.Element {
+  override val key: ExecutionContext.Key<*>
+    get() = Key
+
+  companion object Key : ExecutionContext.Key<DoNotStoreContext>
+}
+
+internal val <D : Operation.Data> ApolloRequest<D>.doNotStore
+  get() = executionContext[DoNotStoreContext]?.value ?: false
+
+/**
+ * @param doNotStore Whether to store the response in cache.
+ *
+ * Default: false
+ */
+fun <T> MutableExecutionOptions<T>.doNotStore(doNotStore: Boolean) = addExecutionContext(
+    DoNotStoreContext(doNotStore)
+)
