+++ normalized-cache/src/commonMain/kotlin/com/apollographql/cache/normalized/options/fetchOnlyIfCached.kt
@@ -1,0 +1,29 @@
+package com.apollographql.cache.normalized.options
+
+import com.apollographql.apollo.api.ExecutionContext
+import com.apollographql.apollo.api.ExecutionOptions
+import com.apollographql.apollo.api.MutableExecutionOptions
+
+internal class FetchOnlyIfCachedContext(val value: Boolean) : ExecutionContext.Element {
+  override val key: ExecutionContext.Key<*>
+    get() = Key
+
+  companion object Key : ExecutionContext.Key<FetchOnlyIfCachedContext>
+}
+
+internal val ExecutionOptions.onlyIfCached: Boolean
+  get() = executionContext[FetchOnlyIfCachedContext]?.value ?: false
+
+/**
+ * Sets whether to only return results from the cache.
+ * If set to true, the network will not be queried at all.
+ *
+ * Default: false
+ */
+internal fun <T> MutableExecutionOptions<T>.onlyIfCached(onlyIfCached: Boolean): T {
+  // noCache and onlyIfCached are mutually exclusive
+  if (onlyIfCached) {
+    addExecutionContext(FetchNoCacheContext(false))
+  }
+  return addExecutionContext(FetchOnlyIfCachedContext(onlyIfCached))
+}
