+++ normalized-cache/src/commonMain/kotlin/com/apollographql/cache/normalized/options/writeToCacheAsynchronously.kt
@@ -1,0 +1,29 @@
+@file:Suppress("PackageDirectoryMismatch")
+
+package com.apollographql.cache.normalized
+
+import com.apollographql.apollo.api.ApolloRequest
+import com.apollographql.apollo.api.ExecutionContext
+import com.apollographql.apollo.api.MutableExecutionOptions
+import com.apollographql.apollo.api.Operation
+
+internal class WriteToCacheAsynchronouslyContext(val value: Boolean) : ExecutionContext.Element {
+  override val key: ExecutionContext.Key<*>
+    get() = Key
+
+  companion object Key : ExecutionContext.Key<WriteToCacheAsynchronouslyContext>
+}
+
+internal val <D : Operation.Data> ApolloRequest<D>.writeToCacheAsynchronously
+  get() = executionContext[WriteToCacheAsynchronouslyContext]?.value ?: false
+
+/**
+ * @param writeToCacheAsynchronously whether to return the response before writing it to the cache
+ *
+ * Setting this to true reduces the latency
+ *
+ * Default: false
+ */
+fun <T> MutableExecutionOptions<T>.writeToCacheAsynchronously(writeToCacheAsynchronously: Boolean) = addExecutionContext(
+    WriteToCacheAsynchronouslyContext(writeToCacheAsynchronously)
+)
